backend/requirements.txt
fastapi[standard]
python-dotenv
SQLAlchemy==1.4.54
snowflake-sqlalchemy
snowflake-connector-python
tabulate
pandas
marshmallow